refactor: CollaboRequest DTO, Entity, Mapper 확인 및 점검
1. DTO 패키지 생성 (Request, Response, Service)
