+++ src/main/java/com/bluehair/hanghaefinalproject/collaboRequest/dto/responseDto/ResponseCollaboRequestListForPostDto.java
@@ -1,4 +1,4 @@
-package com.bluehair.hanghaefinalproject.collaboRequest.dto;
+package com.bluehair.hanghaefinalproject.collaboRequest.dto.responseDto;
 
 import com.bluehair.hanghaefinalproject.common.service.LocalDateTimeConverter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Schema(description = "콜라보 리퀘스트 전체 조회 Dto")
 @Getter
-public class CollaboRequestListForPostDto {
+public class ResponseCollaboRequestListForPostDto {
     private Long collaboId;
     @Schema(description = "닉네임", example = "test01")
     private String nickname;
@@ -31,9 +31,9 @@ public class CollaboRequestListForPostDto {
     private List<String> musicPartsList;
 
     @Builder
-    public CollaboRequestListForPostDto(Long collaboId, String nickname, String contents,
-                                        String profileImg, Long followerCount, Boolean isFollowed, LocalDateTime createdAt,
-                                        LocalDateTime modifiedAt, List<String> musicPartsList) {
+    public ResponseCollaboRequestListForPostDto(Long collaboId, String nickname, String contents,
+                                                String profileImg, Long followerCount, Boolean isFollowed, LocalDateTime createdAt,
+                                                LocalDateTime modifiedAt, List<String> musicPartsList) {
         this.collaboId = collaboId;
         this.nickname = nickname;
         this.contents = contents;
