+++ src/main/java/com/bluehair/hanghaefinalproject/collaboRequest/dto/serviceDto/CollaboRequestDetailsDto.java
@@ -1,4 +1,4 @@
-package com.bluehair.hanghaefinalproject.collaboRequest.dto;
+package com.bluehair.hanghaefinalproject.collaboRequest.dto.serviceDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
